refactor(cli): unify loop run presentation
Separate live dashboard rendering from list and history output, and route
manual and stored run forensics through one report module without changing
rendered bytes.
