+++ crates/rimz/src/cli/loop_cmd/run.rs
@@ -1,15 +1,11 @@
 //! Execute loop tasks and record foreground or scheduled run outcomes.
 
+use super::run_report::{
+    RunSummary, write_check_trip_line, write_manual_verdict, write_run_summary,
+};
 use super::*;
 
-const CHECK_SUMMARY_OUTPUT_CAP: usize = 4 * 1024;
-
 // ---- run --------------------------------------------------------------------
-
-struct RunSummary<'a> {
-    record: &'a LoopRunRecord,
-    presentation: &'a LoopRunPresentation,
-}
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 enum ProjectTrustDecision {
@@ -360,19 +356,6 @@ fn write_manual_header(
     )
 }
 
-fn write_manual_verdict(
-    out: &mut impl Write,
-    result: LoopRunResult,
-    label: &str,
-) -> std::io::Result<()> {
-    let mark = render::loop_result_mark(result);
-    writeln!(
-        out,
-        "{}",
-        ui::paint(mark.style, &format!("{} {label}", mark.glyph))
-    )
-}
-
 fn print_run_summary(
     name: &str,
     entry: &TaskEntry,
@@ -384,391 +367,6 @@ fn print_run_summary(
     let mut out = ui::out();
     write_run_summary(&mut out, name, entry, action, mode, keep, summary)?;
     Ok(())
-}
-
-fn write_run_summary(
-    out: &mut impl Write,
-    name: &str,
-    entry: &TaskEntry,
-    action: &TaskAction,
-    mode: LoopRunMode,
-    keep: bool,
-    summary: &RunSummary<'_>,
-) -> std::io::Result<()> {
-    let action_kind = action.kind();
-    match mode {
-        LoopRunMode::Manual => {
-            write_manual_run_summary(out, name, entry, action, action_kind, keep, summary)
-        }
-        LoopRunMode::Scheduled => write_scheduled_run_summary(out, name, entry, action, summary),
-    }
-}
-
-fn write_manual_run_summary(
-    out: &mut impl Write,
-    name: &str,
-    entry: &TaskEntry,
-    action: &TaskAction,
-    action_kind: TaskActionKind,
-    keep: bool,
-    summary: &RunSummary<'_>,
-) -> std::io::Result<()> {
-    let record = summary.record;
-    let duration_ms = record.duration_ms.unwrap_or_default();
-    if record.result == LoopRunResult::CheckSkipped {
-        return write_check_skipped_summary(
-            out,
-            name,
-            entry,
-            action,
-            duration_ms,
-            LoopRunMode::Manual,
-            summary,
-        );
-    }
-    if let Some((result, label)) = manual_early_verdict(summary) {
-        return write_manual_verdict(out, result, &label);
-    }
-
-    let result_mark = render::loop_result_mark(record.result);
-    let result_label = manual_result_label(action_kind, summary);
-    write!(
-        out,
-        "{}",
-        ui::paint(
-            result_mark.style,
-            &format!("{} {result_label}", result_mark.glyph)
-        )
-    )?;
-    write!(out, " in {}", render::format_duration_ms(duration_ms))?;
-    if let Some(spend) =
-        render::spend_segments(record.cost_usd, record.input_tokens, record.output_tokens)
-    {
-        write!(out, " · {spend}")?;
-    }
-    writeln!(out)?;
-
-    if is_spawn_failure(record.result) && !action_kind.is_check_only() {
-        write_failure_forensics(out, name, summary)?;
-    } else if record.result == LoopRunResult::Completed && record.run_id.is_some() {
-        write_completion_detail(out, name, summary)?;
-    }
-    if !is_spawn_failure(record.result) && !keep && record.run_id.is_some() {
-        writeln!(
-            out,
-            "{}",
-            ui::paint(
-                ui::palette::muted(),
-                "  pane closed; rerun with --keep to watch"
-            )
-        )?;
-    }
-    Ok(())
-}
-
-fn manual_early_verdict(summary: &RunSummary<'_>) -> Option<(LoopRunResult, String)> {
-    let label = match summary.record.result {
-        LoopRunResult::Expired => "deadline expired — task left in place".to_owned(),
-        LoopRunResult::TargetGone => format!(
-            "{} not alive — schedule left in place",
-            summary.record.target.as_deref().unwrap_or("target")
-        ),
-        _ => return None,
-    };
-    Some((summary.record.result, label))
-}
-
-fn is_spawn_failure(result: LoopRunResult) -> bool {
-    matches!(
-        result,
-        LoopRunResult::Failed
-            | LoopRunResult::VerifyFailed
-            | LoopRunResult::TimedOut
-            | LoopRunResult::BudgetExceeded
-    )
-}
-
-fn write_scheduled_run_summary(
-    out: &mut impl Write,
-    name: &str,
-    entry: &TaskEntry,
-    action: &TaskAction,
-    summary: &RunSummary<'_>,
-) -> std::io::Result<()> {
-    let record = summary.record;
-    let duration_ms = record.duration_ms.unwrap_or_default();
-    if record.result == LoopRunResult::CheckSkipped {
-        return write_check_skipped_summary(
-            out,
-            name,
-            entry,
-            action,
-            duration_ms,
-            LoopRunMode::Scheduled,
-            summary,
-        );
-    }
-    let result_mark = render::loop_result_mark(record.result);
-    let exit_label = outcome_exit_label(summary);
-    if is_spawn_failure(record.result) {
-        let mut label = record.result.label().to_owned();
-        if let Some(exit_label) = exit_label.as_deref() {
-            label.push(' ');
-            label.push_str(exit_label);
-        }
-        write!(
-            out,
-            "loop `{name}`: {}",
-            ui::paint(result_mark.style.bold(), &label)
-        )?;
-        write!(out, " in {}", render::format_duration_ms(duration_ms))?;
-        if let Some(spend) =
-            render::spend_segments(record.cost_usd, record.input_tokens, record.output_tokens)
-        {
-            write!(out, " · {spend}")?;
-        }
-        writeln!(out)?;
-        write_failure_forensics(out, name, summary)?;
-    } else {
-        let result_label = success_result_label(record);
-        write!(
-            out,
-            "loop `{name}`: {}",
-            ui::paint(result_mark.style, &result_label)
-        )?;
-        if let Some(exit_label) = exit_label.as_deref() {
-            write!(out, " {exit_label}")?;
-        }
-        write!(out, " in {}", render::format_duration_ms(duration_ms))?;
-        if let Some(spend) =
-            render::spend_segments(record.cost_usd, record.input_tokens, record.output_tokens)
-        {
-            write!(out, " · {spend}")?;
-        }
-        writeln!(out)?;
-        if record.result == LoopRunResult::Completed && record.run_id.is_some() {
-            write_completion_detail(out, name, summary)?;
-        }
-    }
-    Ok(())
-}
-
-fn success_result_label(record: &LoopRunRecord) -> String {
-    match (record.result, record.target.as_deref()) {
-        (LoopRunResult::Delivered, Some(target)) => format!("delivered to {target}"),
-        _ => record.result.label().to_owned(),
-    }
-}
-
-fn manual_result_label(action_kind: TaskActionKind, summary: &RunSummary<'_>) -> String {
-    if action_kind.is_check_only()
-        && let Some(check) = &summary.record.check
-    {
-        return check_result_label(check);
-    }
-    let mut label = success_result_label(summary.record);
-    if let Some(exit_label) = outcome_exit_label(summary) {
-        label.push(' ');
-        label.push_str(&exit_label);
-    }
-    label
-}
-
-fn check_result_label(check: &CheckRecord) -> String {
-    if check.timed_out {
-        "check timed out".to_owned()
-    } else if check.code == Some(0) {
-        "check passed (exit 0)".to_owned()
-    } else {
-        match check.code {
-            Some(code) => format!("check failed (exit {code})"),
-            None => "check failed (signal)".to_owned(),
-        }
-    }
-}
-
-fn write_check_trip_line(
-    out: &mut impl Write,
-    action: &TaskAction,
-    check: &CheckRecord,
-    duration_ms: u64,
-) -> std::io::Result<()> {
-    let (glyph, style) = if check.timed_out || check.code != Some(0) {
-        ("✗", ui::palette::alarm())
-    } else {
-        ("✓", ui::palette::good())
-    };
-    write!(
-        out,
-        "  {}",
-        ui::paint(
-            style,
-            &format!(
-                "{glyph} {} in {}",
-                check_result_label(check),
-                render::format_duration_ms(duration_ms)
-            )
-        )
-    )?;
-    writeln!(
-        out,
-        " {}",
-        ui::paint(
-            ui::palette::accent(),
-            &format!("→ {}", render::action_progressive_phrase(action))
-        )
-    )
-}
-
-fn write_check_skipped_summary(
-    out: &mut impl Write,
-    name: &str,
-    entry: &TaskEntry,
-    action: &TaskAction,
-    duration_ms: u64,
-    mode: LoopRunMode,
-    summary: &RunSummary<'_>,
-) -> std::io::Result<()> {
-    let label = summary
-        .record
-        .check
-        .as_ref()
-        .map(check_result_label)
-        .unwrap_or_else(|| "check skipped".to_owned());
-    let check_duration_ms = summary
-        .presentation
-        .check_duration_ms
-        .unwrap_or(duration_ms);
-    let duration = render::format_duration_ms(check_duration_ms);
-    let (glyph, style) = render::check_skip_display(summary.record.check.as_ref());
-    if mode == LoopRunMode::Manual {
-        write!(
-            out,
-            "{}",
-            ui::paint(style, &format!("{glyph} {label} in {duration}"))
-        )?;
-        writeln!(
-            out,
-            "{}",
-            ui::paint(
-                ui::palette::muted(),
-                &format!(" — {}", render::check_skip_decision(entry, action))
-            )
-        )
-    } else {
-        write!(out, "loop `{name}`: {}", ui::paint(style, &label))?;
-        writeln!(
-            out,
-            " in {duration} — {}",
-            render::check_skip_decision(entry, action)
-        )
-    }
-}
-
-fn write_failure_forensics(
-    out: &mut impl Write,
-    name: &str,
-    summary: &RunSummary<'_>,
-) -> std::io::Result<()> {
-    if let Some(tail) = outcome_failure_tail(summary) {
-        write_failure_tail(out, &tail)?;
-    }
-    write_run_links(out, summary.record)?;
-    writeln!(
-        out,
-        "{}",
-        ui::paint(
-            ui::palette::muted(),
-            &format!("  see: rimz loop show {name}")
-        )
-    )
-}
-
-fn write_run_links(out: &mut impl Write, record: &LoopRunRecord) -> std::io::Result<()> {
-    if let Some(run_id) = &record.run_id {
-        writeln!(
-            out,
-            "{}",
-            ui::paint(ui::palette::muted(), &format!("  run: {run_id}"))
-        )?;
-    }
-    if let Some(transcript) = &record.transcript_path {
-        writeln!(
-            out,
-            "{}",
-            ui::paint(ui::palette::muted(), &format!("  transcript: {transcript}"))
-        )?;
-    }
-    Ok(())
-}
-
-fn write_completion_detail(
-    out: &mut impl Write,
-    name: &str,
-    summary: &RunSummary<'_>,
-) -> std::io::Result<()> {
-    if !summary.presentation.streamed {
-        if let Some(message) = summary
-            .record
-            .last_message
-            .as_deref()
-            .filter(|msg| !msg.trim().is_empty())
-        {
-            render::write_gutter_block(out, None, message)?;
-        } else {
-            writeln!(
-                out,
-                "{}",
-                ui::paint(
-                    ui::palette::muted(),
-                    &format!("  no final message; see: rimz loop show {name}")
-                )
-            )?;
-        }
-    }
-    write_run_links(out, summary.record)
-}
-
-fn outcome_exit_label(summary: &RunSummary<'_>) -> Option<String> {
-    if let Some(exit) = summary.presentation.exit_code {
-        if exit == 0 {
-            return None;
-        }
-        Some(format!("(exit {exit})"))
-    } else if let Some(exit) = summary.record.check.as_ref().and_then(|check| check.code) {
-        Some(format!("(exit {exit})"))
-    } else if summary
-        .record
-        .check
-        .as_ref()
-        .is_some_and(|check| check.timed_out)
-    {
-        Some("(timeout)".to_owned())
-    } else {
-        None
-    }
-}
-
-fn outcome_failure_tail(summary: &RunSummary<'_>) -> Option<String> {
-    if let Some(tail) = summary
-        .presentation
-        .failure_tail
-        .as_deref()
-        .filter(|tail| !tail.trim().is_empty())
-    {
-        return Some(tail.trim_end().to_owned());
-    }
-    let check = summary.record.check.as_ref()?;
-    if !check.timed_out && check.code == Some(0) {
-        return None;
-    }
-    let tail = rimz::proc::tail_output(check.output.as_bytes(), CHECK_SUMMARY_OUTPUT_CAP);
-    let tail = tail.trim_end();
-    (!tail.trim().is_empty()).then(|| tail.to_owned())
-}
-
-fn write_failure_tail(out: &mut impl Write, tail: &str) -> std::io::Result<()> {
-    render::write_gutter_block(out, Some(ui::palette::alarm()), tail)
 }
 
 #[cfg(test)]
